refactor : toolbar의 긴 코드를 함수로 나눠서 보기 편하게 적용
