+++ src/components/toolbar/utils/findDivParent.js
@@ -1,0 +1,9 @@
+export default function findDivParent($element){
+  let $targetElement = $element;
+
+  while($targetElement.nodeName !== "DIV"){
+    $targetElement = $targetElement.parentElement;
+  }
+
+  return $targetElement;
+}
