+++ src/utils/api.js
@@ -4,7 +4,6 @@ import { replaceHistory } from "./router.js";
 export const request = async (url, option = {}) => {
   try{
     const res = await fetch(`${API_END_POINT}${url}`, {
-      ...PushSubscriptionOptions,
       ...option,
       headers : {
         "Content-Type" : "application/json",
